Revert fix for 874:
Unfortunately, there is no public API in the VS debugger infrastructure to either ask to refresh break state (locals, etc.) without affecting focus, or to get an event when break state handling, including focus handling, is complete. This leaves our solution to keeping break state (locals, etc.) up to date with changes in the debug REPL with the nasty side effect of automatically changing focus to the editor window. These changes back out this original fix, as the bad side effect outweighs the improvement of keeping break state in sync.