1817 from ... import completions broken on PyQt Ensures child modules are also added as members when saving the database. Adds test.
1819 Database refreshes strangely after PyQt install
Ensures that when '_all' is changed to true, all the preceding file groups are included as well as subsequent ones.

Also fixes extension modules not always being importable when the actual sys.path does not match what we believe it does.
Ensures the correct analysis limits are loaded.
Moves registry settings for analysis limits to the Microsoft\PythonTools key rather than the Microsoft\VisualStudio\1x.0\PythonTools key. Because they are often loaded from outside of VS, it makes more sense to keep them outside of the VS hive.