#2050 Django template editing crash
Do not force HTML editor tree to be up-to-date when classifying. This may cause queued editing actions (like tag autocomplete) to be applied to the text buffer, and doing so inside GetClassificationSpans breaks the editor. Instead, if the tree is not up to date, don't classify anything, but asynchronously notify the editor that it needs to re-classify once the tree catches up with the buffer.

Various unrelated unit test fixes to get 100% pass rate on DjangoEditingTests.