#2085 Locals window treats attributes referenced in current frame as global variables
The code was mistakenly assuming that co_names in a local frame only refer to locals and globals, and so any name that was not a local should be treated as a global, and displayed as <unknown> if it it not actually in f_globals. In practice, co_names lists all names referenced by a given frame, including e.g. attribute names. The correct behavior is achieved by only displaying those co_names which have a matching entry in f_globals (false positives can still happen, but are much rarer now). This also means that we will no longer display unassigned-but-referenced globals as <unknown>.