Fixes bugs & implements feature requests that are to do with enumerating object children:
#2019 Instance of class with __call__ or __getattr__ method not shown when attribute of another instance
#2065 What items should be shown in the VS Locals window?
#2074 Bug: Locals Window does not expand or show a dictionary item correctly.
#2090 No way to expand dict keys
#2091 Add the ability to evaluate and expand generators

The major change underscoring these is how child items are handled now. Where previously the Python debugger script only reported the child name relative to its parent, and whether it's an attr (.abc) or an index (['abc']), and on VS side a full expression was reconstituted from that and the parent expression, now the debugger script computes the full expression and reports it to VS, which uses it as is. This considerably simplifies the protocol and reduces the number of code paths when handling children, and also allows for new child types with arbitrary expressions.

This, in turn, is used to rewrite the code handling non-indexable sequences such as sets. For those, we now generate the full expression for child elements using enumerate() and a sequence comprehension - this makes it possible to use Add Watch for set elements. Furthermore, the non-indexability check is now also applied to dicts, to gracefully handle dict keys which do not have an eval'able repr.

This is also used to add a synthetic child node representing items() (viewitems() in 2.7) for a given dict. As it represents dict as a sequence of key-value tuples, it permits inspecting keys as full-fledged objects. A similar synthetic node called Results View is added to generators, and can be used to force their iteration and obtain the result as a tuple that can then be inspected.

The logic used to filter out attributes representing methods is also different now. We now display all attributes that are found on the instance itself regardless of type, even if callable (so an attribute referencing some function is no longer hidden). For attributes that come from the class of the instance, we filter out methods using a hardcoded list of types, rather than anything with a __call__ - so e.g. nested classes are not filtered out.