#2211 Support remote debugging of Azure web sites
Add ptvsd support to ptvs_virtualenv_proxy.py, triggered by the presence of WSGI_PTVSD_SECRET environment variable / appSettings key.

Split common code in WebSocketProxy in a separate base class and share it between PTVS and NTVS.

Include WebRole.dll in the list of files to publish to Azure, and auto-generate web.debug.config transform that exposes it as an endpoint on the server.

Change remote attach qualifiers to be generic URLs (with a fallback old-style syntax), just like NTVS. This also removes the need for two separate transports for unsecure and SSL connections, as this is now distinguished by the URL scheme (tcp:// vs tcps://). Refactor connection handling code to be cleanly separated from the rest of remote debugging logic.

Change debugger and REPL code to avoid assuming that it always has a valid raw socket to debuggee - the connection is now just a generic stream, and a socket is only used for the listen-for-incoming-connection case, and only until connection is established. There are some other minor unrelated refactorings thrown into this to simplify the code in general.

Misc associated test fixes.