#2375 Stop debugging before hitting any breakpoint gives error, VS hang Raise IDebugProgramCreateEvent2 as soon as we attach, rather than waiting for code to execute. This way, debugger doesn't think that we're still attaching if user tries to stop immediately after attaching and before any Python code had a chance to run.
#2398 Error Current interactive window is disconnected in Python Debug Inteactive Window
If debugger connection is dropped for any reason, treat it as the end of the debug session and raise the corresponding event. This prevents debugger and REPL from being stuck in a phantom state where they think they're still debugging while the debuggee is gone (e.g. because it was forcibly killed).

#2372 Attach debugger to azure web site error when sockets are disabled
Add a special error message for HTTP 503 status that directs the user to check websocket setting in Azure portal.

#2387 Enable attach debugger via web sockets on VS 2010
Remove #if for WebSocketStream and the corresponding transport, since we're using .NET 4.5, and websockets are available regardless of VS version. This does not enable Server Explorer integration on VS 2010, but only wss:// URLs in Debug -> Attach to Process.

#2373 Ping Azure website over HTTP before trying to attach to it
To make sure that the server process is running, and there is something to attach to.

#2313 Debugger crash
This is just hardening of external input rather than an actual fix, since the root cause couldn't be determined at the point of the crash. Instead of crashing on an unknown evaluation result, Debug.Assert and ignore it.

#2174 Tools->Options->Text Editor->DjangoTemplate should have a space